# Conflicts:
#	backend/Widget.Api/ApiModels/PostEventApiModel.cs

+++ backend/Widget.Api/ApiModels/PostEventApiModel.cs
@@ -1,4 +1,4 @@
 ﻿namespace Widget.Api.ApiModels;
 
 public record PostEventApiModel(EventType Event, string IssueId, string Login, string ProjectKey, string ProjectName, 
-    string[] Children, Priority IssuePriority, ulong StoryPoints);
+    string[] Children, Priority IssuePriority, uint StoryPoints);
